requirements.txt
playwright==1.58.0
requests
python-dotenv
beautifulsoup4
apscheduler